+++ Kokoro/APIv2/StatusLogic.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
 namespace Shockah.Kokoro;
@@ -97,6 +98,14 @@ public partial interface IKokoroApi
 				bool? IsAffectedByBoost(IIsAffectedByBoostArgs args) => null;
 
 				/// <summary>
+				/// Controls the statuses for which the <see cref="ModifyStatusTurnTriggerPriority"/>, <see cref="OnStatusTurnTrigger"/> and <see cref="HandleStatusTurnAutoStep"/> methods will be called.
+				/// Defaults to <c>args.NonZeroStatuses</c>.
+				/// </summary>
+				/// <param name="args">The arguments for the hook method.</param>
+				/// <returns>The set of statuses other hooks will be called for.</returns>
+				IReadOnlySet<Status> GetStatusesToCallTurnTriggerHooksFor(IGetStatusesToCallTurnTriggerHooksForArgs args) => args.NonZeroStatuses;
+
+				/// <summary>
 				/// Modifies the priority at which a status triggers at the start and end of every turn.
 				/// </summary>
 				/// <param name="args">The arguments for the hook method.</param>
@@ -177,6 +186,42 @@ public partial interface IKokoroApi
 					/// The status being changed.
 					/// </summary>
 					Status Status { get; }
+				}
+
+				/// <summary>
+				/// The arguments for the <see cref="IHook.GetStatusesToCallTurnTriggerHooksFor"/> hook method.
+				/// </summary>
+				public interface IGetStatusesToCallTurnTriggerHooksForArgs
+				{
+					/// <summary>
+					/// The game state.
+					/// </summary>
+					State State { get; }
+					
+					/// <summary>
+					/// The current combat.
+					/// </summary>
+					Combat Combat { get; }
+
+					/// <summary>
+					/// The current timing of the turn.
+					/// </summary>
+					StatusTurnTriggerTiming Timing { get; }
+
+					/// <summary>
+					/// The ship the status is getting triggered for.
+					/// </summary>
+					Ship Ship { get; }
+					
+					/// <summary>
+					/// All existing statuses in the game.
+					/// </summary>
+					IReadOnlySet<Status> KnownStatuses { get; }
+					
+					/// <summary>
+					/// All statuses the ship currently has a non-zero amount of.
+					/// </summary>
+					IReadOnlySet<Status> NonZeroStatuses { get; }
 				}
 
 				/// <summary>